9397: Move refableQings fetcher into FiltersModel
Now it updates when the selected forms change.
